requirements.txt
functions-framework==2.2.1
flask==2.0.1
google-cloud-error-reporting==1.1.2
google-cloud-firestore==2.1.3